refactor: 지도 관련 파일 구조 정리 (컴포넌트/훅/유틸 분리)
